// SMSfrontEnd/src/app/models/item-service.ts
import { Category } from "./category";

export class ItemService {
    id: number;
    itemServiceName: string;
    code: string;
    uom: string;
    // categoryName: string;
    description: string;   
    purchasedPrice: number;
    tax: number;
    sellingUnitPrice: number;
    quantity: number;  
    categoryId: any;
    // category : Category;


    //   constructor(category : Category) {
    //         this.category = category;
    //      }
        
   
}